refactor: Extract useDOMMutations to own file
